Decouple HashCodeBuilder from Builder
We are deprecating Builder concept and HashCodeBuilder does not really
like having to inflate to java.lang.Integer. Ditch the indirection.

JIRA: YANGTOOLS-1327

+++ common/util/src/test/java/org/opendaylight/yangtools/util/HashCodeBuilderTest.java
@@ -12,16 +12,15 @@ import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 
 public class HashCodeBuilderTest {
-
     @Test
     public void testAllMethodsOfHashCodeBuilder() {
         final HashCodeBuilder<String> builder = new HashCodeBuilder<>();
-        assertEquals("Default hash code should be '1'.", 1, builder.build().intValue());
+        assertEquals("Default hash code should be '1'.", 1, builder.build());
 
         int nextHashCode = HashCodeBuilder.nextHashCode(1, "test");
         assertEquals("Next hash code should be '3556529'.", 3556529, nextHashCode);
 
         builder.addArgument("another test");
-        assertEquals("Updated internal hash code should be '700442706'.", -700442706, builder.build().intValue());
+        assertEquals("Updated internal hash code should be '700442706'.", -700442706, builder.build());
     }
 }
